+++ test/setup/makeTestSettings.ts
@@ -9,7 +9,7 @@ import TestSettings, {
 	TestRunnerSettings,
 	TestSuiteSettings
 } from '../types/TestSettings.js';
-import loadEnvFiles from './loadEnvFiles.js';
+import loadEnvFiles from './load-env-files.js';
 
 export const ONE_DAY_IN_MS = 86400000;
